Refactoring 10 (#59)
* Change release builds from RelWithDebInfo back to Release

I entirely forgot that RelWithDebugInfo is pretty much a debug build.

* Make more data const

* Make more data const

And re-order a global to avoid it taking up space in the data section.

* Remove unnecessary zero initialization

C++ (and C) will already zero initialize so remove the explicit just to be clear what we are actually initializing.

* Make more globals data const

* Reformatting

Compilation proves no change

* Tidy up of more globals

Remove unreferenced, make static if it can be, make const if it can be.

* Change signature of enchant_weapon to take item

No longer rely on global state

* Move enchant_weapon into common library

* Reformat equip_item

Checked before/after for same executable

* Remove dependency on global adven

* Pull out helper function ahead of move to common library

* Extract pc_could_accept function to common library

* Extract helper function

* Remove global state reference from helper function

* Move pc_remove_item to common library

I accept that returning bool to indicate the loss of poison weaponed status is poor, but I've yet to conclude a strategy for that sort of thing.

* Move return_treasure to common library

* Cleanups

* Some more tidy up to use ranged for loops where possible

* Change cd_get_text_edit_str to return a std::string

* Address https://github.com/CorrodedCoder/Blades-of-Exile/issues/58

Compare only the low word of `wparam` in comparisons against the id of the control and compare the high word of `wparam` against the `EN_ERRSPACE` notification rather than the high word of `lparam`.

* Cut out the middle man and avoid the c_str() member calls

Leverage std::stoi instead of sscanf to safely convert strings to integer values.

* Replace another sscanf with std::stoi

* Reformat wd_to_pwd

* Correct bug I introduced by changing character array param to std::string_view

* Change long type to int

As per a previous issue, for a 16-bit app a long was 32-bit, but for 64-bit it will now be 64-bit, so change it to an int.

* Add exception catch for failed ascii to int conversion

Also tidy up wd_to_pwd

* Move registration scenario password  algorithm to new registration component of common library

* Remove unused dummy parameter to Delay

Also tidy up clang warnings to registration

* Move Delay function to common library

Implementation changes from a manual check to just relying on the c++ std::this_thread::sleep_for API.

* Replace sound_pause with direct call to sleep 25 milliseconds

* Move same_point to common library

* make same_point return bool and add not_same_point

* Change remaining `long` game types to `int`

The game was written assuming that long was 32-bit, but on 64-bit platforms it is not always so (depending on whether it's an LP64 compiler or just a P64 compiler like MSVC).
The remaining untouched longs are actually being used to track time duration so they will eventually change to a more suitable type.

* Remove more `long` types

Some weren't referenced, some were set but not referenced, and some needed to be changed to DWORD type.

* Change file offsets from long to size_t

+++ src/Windows Code Release 3/fileio.h
@@ -16,7 +16,7 @@ void fix_boats();
 void load_outdoors(short to_create_x, short to_create_y, short targ_x, short targ_y,
 	short mode,short extra,char *str);
 void get_reg_data();
-long do_waterfall(long flag);
+int do_waterfall(int flag);
 short init_data(short flag);
 void start_data_dump();
 void end_data_dump();
